+++ RELEASE_NOTES.md
@@ -1,7 +1,7 @@
-﻿# Hotspot Manager v0.5.22
+﻿# Hotspot Manager v0.5.23
 
 ## Yenilikler
 
 
 ## Notlar
-- Fixed Access Denied error on startup toggle by adding UAC elevation prompt
+- Fixed startup issues by ensuring Task Scheduler power settings allow execution on battery and setting correct Working Directory